change client id related type from int to uintptr_t to prevent truncation

+++ src/GameClient.cpp
@@ -19,7 +19,7 @@ void GameClient::onMessageFromServer(Message& message)
 }
 
 
-int GameClient::getClientID()
+uintptr_t GameClient::getClientID()
 {
     return m_clientID;
 }
